# fix: avoid alloc using strings.IndexFunc instead